% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcEntropy}
\alias{calcEntropy}
\title{Calculate the entropy for the current cluster weights.}
\usage{
calcEntropy(class_weights)
}
\arguments{
\item{class_weights}{The vector of the cluster weights from a mixture model.}
}
\value{
The information entropy for the current cluster weights.
}
\description{
Calculate the entropy for the current cluster weights.
}
