% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_methods.R
\name{categorical_gibbs_sampling}
\alias{categorical_gibbs_sampling}
\title{Categorical gibbs sampling}
\usage{
categorical_gibbs_sampling(data, fix_vec = rep(F, nrow(data)),
  d = ncol(data), N = nrow(data),
  cluster_weight_priors_categorical = 1, phi_0 = NULL,
  c_clusters_label_0 = NULL, num_clusters_cat = NULL,
  num_iter = 10000, burn = floor(num_iter/10), thinning = 25)
}
\arguments{
\item{data}{A matrix of the data being analysed.}

\item{cluster_weight_priors_categorical}{Vector of the prior on cluster
weights in the categorical data.}

\item{phi_0}{List of vectors, the prior on the distribution of the classes
over clusters.}

\item{c_clusters_label_0}{Vector of labels for the prior clustering of the
categorical data.}

\item{num_clusters_cat}{Integer of the number of clusters to have as a
maximum in the categorical dataset. Default is 100.}

\item{num_iter}{The number of iterations to sample over.}

\item{burn}{The number of iterations to record after (i.e. the burn-in).}

\item{thinning}{The step between iterations for which results are recorded in
the mcmc output.}
}
\description{
Carries out gibbs sampling of data and returns a similarity matrix for points
}
