% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{checkThinningFactor}
\alias{checkThinningFactor}
\title{Check Thinning Factor}
\usage{
checkThinningFactor(thinning, num_iter, burn)
}
\value{
None.
}
\description{
Generates an error or warning if the relative values of
thinning, num_iter and burn (the arguments controlling the number of recorded
MCMC samples) are strange.
}
