% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countCategories}
\alias{countCategories}
\title{Returns a vector of the number of unqiue values in each column of data.}
\usage{
countCategories(data)
}
\arguments{
\item{data}{Matrix of categorical data.}
}
\value{
Vector of counts of the number of categories in each variable in data.
}
\description{
Returns a vector of the number of unqiue values in each column of data.
}
