% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_creation.R
\name{create_dataset_folders}
\alias{create_dataset_folders}
\title{Create dataset folders}
\usage{
create_dataset_folders(n_clust, type, dataset_num, overwrite = FALSE)
}
\arguments{
\item{n_clust}{Integer. The number of clusters in the dataset.}

\item{type}{String. The type of muixture model used, Gaussian or Categorical.}

\item{dataset_num}{Integer. The number identifying this dataset.}

\item{overwrite}{Bool. Instructs overwriting previously existing directories.}
}
\description{
Creates the directories to hold the dataset specific outputs
(i.e. parameter posteriors and allocation probabilities)
}
