% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_creation.R
\name{create_folder}
\alias{create_folder}
\title{Create folder}
\usage{
create_folder(folder_name, root_dir = ".", overwrite = FALSE)
}
\arguments{
\item{folder_name}{A string. The name for the folder.}

\item{root_dir}{A string. The name of the root directory to create the folder
in. Default is the current working directory (i.e. ".").}

\item{overwrite}{A bool. Instructs overwriting if the desired directory is
found to already exist. Default is FALSE (in which case an error is thrown
if the directory does already exist).}
}
\description{
Creates a folder in the appropriate location (default is the
current directory) with an option to overwrite.
}
