% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{createGaussianArguments}
\alias{createGaussianArguments}
\title{Create Gaussian Arguments}
\usage{
createGaussianArguments(
  data,
  n_clust,
  mu_0 = NULL,
  scale_0 = NULL,
  lambda_0 = NULL,
  df_0 = NULL
)
}
\arguments{
\item{data}{A matrix of continuous values.}

\item{n_clust}{A unsigned integer; the number of clusters to be used in the
clustering method.}

\item{mu_0}{A d-vector where d is the number of columns in data representing
the prior beliefs about the mean for data. Defaults to the mean of data if
given value is NULL.}

\item{scale_0}{A positive definite matrix. The prior for the scale parameter
of the inverse wishart distribution; if NULL defaults to a diagonal matrix.}

\item{lambda_0}{A positive real number; the shrinkage prior for the mean.}

\item{df_0}{An integer. The degrees of freedom used in the inverse Wishart
distribution from which the variance is sampled. If NULL defaults to d + 2.}
}
\value{
A named list ready to be used as input into the MDI function.
}
\description{
Creates a named list of the priors required for a mixture of
Gaussians. If the user provides values the function will wrap them in a
single object appropriate for input into the MDI function. If values are not
given the priors are generated using an empirical method (hence the need for
inputting the data and the number of clusters).
}
\examples{
args <- CreateGaussianArguments(data, 5)
}
