% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CreateSimilarityMat}
\alias{CreateSimilarityMat}
\title{Constructs a similarity matrix comparing all points clustering across the 
iterations.}
\usage{
CreateSimilarityMat(cluster_record)
}
\arguments{
\item{cluster_record}{Matrix of label assignment for data across iterations.}
}
\value{
A symmetric n x n matrix (for n rows in cluster record) describing 
the fraction of iterations for which each pairwise combination of points are
assigned the same label.
}
\description{
Constructs a similarity matrix comparing all points clustering across the 
iterations.
}
