% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{crossValidateMDI}
\alias{crossValidateMDI}
\title{Cross validate MDI}
\usage{
crossValidateMDI(
  MS_object,
  MS_cat_object = NULL,
  times = 10,
  test_size = 0.2,
  num_iter = 1000,
  burn = floor(num_iter/10),
  thinning = 25,
  n_clust_cat = 50,
  ...
)
}
\arguments{
\item{MS_object}{A dataset in the format used by pRolocdata.}

\item{MS_cat_object}{A second dataset in the format used by pRolocdata.}

\item{times}{An integer; the number of folds to run.}

\item{test_size}{A fraction between 0 and 1; the proportion of the known
points to use as a test set in each fold.}

\item{num_iter}{An integer; the number of iterations to use in each MDI
sampling.}

\item{burn}{The number of iterations to record after (i.e. the burn-in).}

\item{thinning}{The step between iterations for which results are recorded in
the mcmc output.}

\item{...}{Any other inputs applicable to mcmc_out.}
}
\value{
A named list of the confusion matrix for allocaitons across each fold
and the associated quadratic loss.
}
\description{
Returns the allocation confusion matrix and quadratic loss score
of MDI applied to the expression data from MS object and MS cat object.
}
\examples{
cv_metrics <- CrossValidateMDI(tan2009r1, tan2009r1goCC)
}
