% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_clustering.R
\name{declare_cluster_weights}
\alias{declare_cluster_weights}
\title{Declare cluster weights}
\usage{
declare_cluster_weights(fix_vec, clust_labels, n_clust, weight_0 = NULL)
}
\arguments{
\item{fix_vec}{Vector of 1's and 0's indicating if the corresponding entry in
the data is a fixed label}

\item{clust_labels}{A vector of unsigned integers representing the current
labelling of points in the data}

\item{n_clust}{A unsigned integer; the number of clusters to be used in the
clustering method.}

\item{weight_0}{One of an int or a p-vector where p = n_clust. If an int
defaults to a p-vector of p repetitions of said int.}
}
\value{
A vector of values to be used as the mass parameters for the weights
for the clusters.
}
\description{
Creates a vector for the mass parameter for cluster weights.
}
