% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{declareClassProbsField}
\alias{declareClassProbsField}
\title{Finds the number of categories in each covariate and declares the appropriate
matrix to record the associated probabilties for each cluster.}
\usage{
declareClassProbsField(cat_count, n_col, n_clust)
}
\arguments{
\item{cat_count}{Number of categories present in each variables.}

\item{n_col}{Number of variables.}

\item{n_clust}{Number of clusters present.}
}
\value{
Field of matrices (must be a field as matrices are not always the
same dimension), where each matrix is associated with a single variable and
holds the probabilities of each category within the variable for the n_clust
clusters.
}
\description{
Finds the number of categories in each covariate and declares the appropriate
matrix to record the associated probabilties for each cluster.
}
