% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{generateClusterLabelPrior}
\alias{generateClusterLabelPrior}
\title{Generate Cluster Label Prior}
\usage{
generateClusterLabelPrior(class_labels_0, train, MS_object, k, N)
}
\arguments{
\item{class_labels_0}{An optional prior for clusters in MS_object. If NULL
defaults to a randomly generated set using the proportion in the labelled
data.}

\item{train}{A NULL, TRUE or FALSE value informing if supervised (TRUE),
semi-supervised (NULL) or unsupervised (FALSE)}

\item{MS_object}{The MS object from pRolocData being analysed}

\item{k}{The number of clusters in the dataset}

\item{N}{The number of observations}

\item{nk}{Table of the frequency of the classes in the datset}
}
\value{
A vector of integers corresponding to the cluster allocation of the N
observations
}
\description{
Generates a vector of labels if required
}
