% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{generateClusterWeightPrior}
\alias{generateClusterWeightPrior}
\title{Cluster weight prior}
\usage{
generateClusterWeightPrior(train_data, outlier = FALSE)
}
\description{
Produces a vector of prior weights for clusters
}
