% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{generateGaussianEmpiricalBayesPrior}
\alias{generateGaussianEmpiricalBayesPrior}
\title{Generate Gaussian Empirical Bayes Prior
Generates priors for the mean, degrees of freedom and scale parameters if not
set.}
\usage{
generateGaussianEmpiricalBayesPrior(
  data,
  mu_0,
  df_0,
  scale_0,
  N,
  k,
  d,
  lambda_0 = 0.01
)
}
\arguments{
\item{data}{A matrix of the data being analysed.}

\item{mu_0}{A d-vector. If NULL defaults to a vector of column means of data.}

\item{df_0}{An integer. If NULL defaults to d + 2.}

\item{scale_0}{A positive definite matrix. The prior for the scale parameter
of the inverse wishart distribution; if NULL defaults to a diagonal matrix.}

\item{N}{The number of entries in data.}

\item{k}{The number of clusters used.}

\item{d}{The number of columns in data.}

\item{lambda_0}{A positive real number; the shrinkage prior for the mean.}
}
\value{
A named list of the three hyperparameters, mean, scale and degrees of
 freedom
 @export
}
\description{
Generate Gaussian Empirical Bayes Prior
Generates priors for the mean, degrees of freedom and scale parameters if not
set.
}
