% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_methods.R
\name{mdi}
\alias{mdi}
\title{MDI gibbs sampling}
\usage{
mdi(data_1, data_2, args_1 = NULL, args_2 = NULL,
  type_1 = "Gaussian", type_2 = "Categorical", n_clust_1 = 50,
  n_clust_2 = 50, labels_0_1 = NULL, labels_0_2 = NULL,
  d = ncol(data_1), N = nrow(data_1), a_0 = 1, b_0 = 0.2,
  fix_vec_1 = rep(0, nrow(data_1)), fix_vec_2 = rep(0, nrow(data_2)),
  cluster_weight_0_1 = NULL, cluster_weight_0_2 = NULL,
  num_iter = NULL, burn = floor(num_iter/10), thinning = 25,
  outlier_1 = FALSE, t_df_1 = 4, normalise_1 = FALSE,
  outlier_2 = FALSE, t_df_2 = 4, normalise_2 = FALSE,
  record_posteriors = FALSE, save_results = FALSE)
}
\arguments{
\item{data_1}{A matrix of the data being analysed (context 1).}

\item{data_2}{A matrix of the data being analysed (context 2).}

\item{args_1}{A named list. The output of either Gaussian arguments or
Categorical arguments dpending on the type of data_1, containing the relevant
priors.}

\item{args_2}{A named list similar to args_1 corrected for the relevant type
of data_2.}

\item{type_1}{A string, one of "Gaussian" or "Categorical" instructing the
function which MDI comparison to use for data_1.}

\item{type_2}{A string similar to type_1, one of "Gaussian" or "Categorical"
instructing the function which MDI comparison to use for data_2.}

\item{n_clust_1}{The number of clusters to use for data_1.}

\item{n_clust_2}{The number of clusters to use for data_2.}

\item{labels_0_1}{A vector of unsigned integers representing the initial
cluster of the corresponding point in data_1.}

\item{labels_0_2}{A vector of unsigned integers representing the initial
cluster of the corresponding point in data_2.}

\item{d}{The number of columns in the data (if not input calculates this).}

\item{N}{The number of observations in the data (defaults to the number of
rows in the data).}

\item{a_0}{The prior shape for the context similarity parameter.}

\item{b_0}{The prior rate for the context similarity parameter.}

\item{fix_vec_1}{A vector of 1's and 0's or else bools used for
semi-supervised data (use all FALSE or 0 if using an unsupervised case).}

\item{fix_vec_2}{A vector of the same type and length as fix_vec but used
for the categorical data - defaults to a vector of FALSEs of length N.}

\item{cluster_weight_0_1}{The prior for dirichlet distribution of cluster
weights.}

\item{cluster_weight_0_2}{Vector of the prior on cluster
weights in the categorical data.}

\item{num_iter}{The number of iterations to sample over.}

\item{burn}{The number of iterations to record after (i.e. the burn-in).}

\item{thinning}{The step between iterations for which results are recorded in
the mcmc output.}

\item{outlier_1}{A bool instructing the sampler to consider an additional
outlier cluster following a t-distribution for data_1 (only relevant if
type_1 is "Gaussian").}

\item{t_df_1}{The degrees of freedom for the outlier t-distribution (default
is 4).}

\item{outlier_2}{A bool instructing the sampler to consider an additional
outlier cluster following a t-distribution for data_2 (only relevant if
type_2 is "Gaussian").}

\item{t_df_2}{The degrees of freedom for the outlier t-distribution (default
is 4).}

\item{record_posteriors}{A bool instructing the mcmc function to record the
posterior distributions of the mean and variance for each cluster
(default is FALSE)}
}
\value{
A named list of the relevant outputs from the MDI MCMC.
}
\description{
Carries out gibbs sampling of data and returns a similarity
matrix for points.
}
