% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_clustering.R
\name{mdi_gauss_cat_clustering}
\alias{mdi_gauss_cat_clustering}
\title{MDI gibbs sampling}
\usage{
mdi_gauss_cat_clustering(data, cat_data, k, class_labels, fix_vec,
  d = NULL, N = NULL, fix_vec_cat = NULL, num_iter = NULL,
  burn = NULL, thinning = 25, mu_0 = NULL, df_0 = NULL,
  scale_0 = NULL, lambda_0 = 0.01, concentration_0 = 0.1, a_0 = 1,
  b_0 = 1, cluster_weight_priors_categorical = 1, phi_0 = NULL,
  c_clusters_label_0 = NULL, num_clusters_cat = NULL,
  outlier = FALSE, t_df = 4, record_posteriors = FALSE,
  normalise = FALSE)
}
\arguments{
\item{data}{A matrix of the data being analysed.}

\item{cat_data}{Matrix of 1's and 0's used for multiple dataset integration.}

\item{k}{The number of clusters.}

\item{class_labels}{A vector of unsigned integers representing the initial
cluster of the corresponding point in data.}

\item{fix_vec}{A vector of 1's and 0's or else bools used for semi-supervised
data (use all FALSE or 0 if using an unsupervised case).}

\item{d}{The number of columns in the data (if not input calculates this).}

\item{N}{The number of observations in the data (defaults to the number of
rows in the data).}

\item{fix_vec_cat}{A vector of the same type and length as fix_vec but used
for the categorical data - defaults to a vector of FALSEs of length N.}

\item{num_iter}{The number of iterations to sample over.}

\item{burn}{The number of iterations to record after (i.e. the burn-in).}

\item{thinning}{The step between iterations for which results are recorded in
the mcmc output.}

\item{mu_0}{A d-vector; prior on mean. If NULL defaults to mean of data.}

\item{df_0}{The prior on the degrees of freedom. if NULL defaults to d + 2.}

\item{scale_0}{The prior on the scale for the Inverse Wishart. If NULL
generated using an empirical method.}

\item{lambda_0}{The prior of shrinkage for mean distribution.}

\item{concentration_0}{The prior for dirichlet distribution of cluster
weights.}

\item{a_0}{The prior shape for the context similarity parameter}

\item{b_0}{The prior rate for the context similarity parameter}

\item{cluster_weight_priors_categorical}{Vector of the prior on cluster
weights in the categorical data.}

\item{phi_0}{List of vectors, the prior on the distribution of the classes
over clusters.}

\item{c_clusters_label_0}{Vector of labels for the prior clustering of the
categorical data.}

\item{num_clusters_cat}{Integer of the number of clusters to have as a
maximum in the categorical dataset. Default is 100.}

\item{outlier}{A bool instructing the sampler to consider an additional
outlier cluster following a t-distribution}

\item{t_df}{The degrees of freedom for the outlier t-distribution (default
is 4)}

\item{record_posteriors}{A bool instructing the mcmc function to record the
posterior distributions of the mean and variance for each cluster
(default is FALSE)}
}
\description{
Carries out gibbs sampling of data and returns a similarity
matrix for points.
}
