% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mdiGaussCat}
\alias{mdiGaussCat}
\title{MDI clustering for a gaussian and cateogrical dataset}
\usage{
mdiGaussCat(
  cont_data,
  cat_data,
  mu_0,
  lambda_0,
  scale_0,
  nu_0,
  a_0,
  b_0,
  clust_wgt_priors_gauss,
  clust_wgt_priors_cat,
  phi_prior,
  clust_labels_gauss,
  clust_labels_cat,
  n_clust_gauss,
  n_clust_cat,
  fix_vec_1,
  fix_vec_2,
  n_iter,
  burn,
  thin,
  allow_outliers = FALSE,
  t_df = 4,
  normalise = FALSE,
  u_1 = 2,
  v_1 = 10,
  rate_gauss_0 = 1L,
  rate_cat_0 = 1L
)
}
\description{
Performs MDI with a mixture of Gaussian and Categorical 
distributions on each dataset.
}
