% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{MS_dataset}
\alias{MS_dataset}
\title{MS dataset}
\usage{
MS_dataset(MS_object, train = NULL)
}
\arguments{
\item{MS_object}{A dataset in the format used by pRolocdata.}

\item{train:}{instruction to include all data (NULL), labelled data (TRUE) or
unlabelled data (FALSE). Default is NULL.}
}
\description{
Given an MS object from pRolocData, returns the numerical data,
the count of each class and the vector recording which labels are fixed (i.e
from the training set) and unfixed.
}
\examples{
data("hyperLOPIT2015") # MS object from pRolocData
MS_data <- MS_dataset(hyperLOPIT2015)
}
