% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_creation.R
\name{output_folders}
\alias{output_folders}
\title{Output folders}
\usage{
output_folders(n_clust_1, n_clust_2, save_results = TRUE,
  overwrite = FALSE, type_1 = "Gaussian", type_2 = "Categorical")
}
\arguments{
\item{n_clust_1}{An integer. The number of clusters in the first dataset.}

\item{n_clust_2}{An integer. The number of clusters in the second dataset.}

\item{save_results}{Bool instructing program to save results to file. Default
is FALSE.}

\item{overwrite}{Bool instructing program to overwrite pre-existing results
if they exist in the current directroy. Default is FALSE with warnings anyway.}

\item{type_1}{String indicating the type of mixture used in dataset 1. This
determines the folder for the appropriate parameters (i.e. if Gaussian
instructs creation of folders for a mean posterior and a variance posterior,
and a class probability posterior for "Categorical"). Default is "Gaussian".}

\item{type_2}{String indicating the type of mixture used in dataset 2.
Default is "Categorical".}
}
\description{
Creates all the relevant directories to save outputs from the
sampling
}
