% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_clustering.R
\name{pca_ms_obj}
\alias{pca_ms_obj}
\title{PCA MS object}
\usage{
pca_ms_obj(MS_object, test_pred, ellipses = FALSE)
}
\arguments{
\item{MS_object}{A dataset from pRolocdata.}

\item{test_pred}{A vector of allocations.}

\item{ellipses}{A bool indicating if clusters should be contained within 
appropriately coloured ellipses on the plot.}

\item{alpha.ind}{A number or vector of numbers between 0 and 1 inddicating 
the transparency of points (1 indicates opaque, 0 transparent).}
}
\value{
PCA plot
}
\description{
Creates a PCA plot of a given MS object from pRolocdata grouping
based on a vector of integers.
}
