% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{pheatmap_cluster_by_col}
\alias{pheatmap_cluster_by_col}
\title{Pheatmap cluster by col}
\usage{
pheatmap_cluster_by_col(num_data, annotation_row, sort_col,
  main = "sense_check", use_col_gaps = TRUE, ...)
}
\arguments{
\item{num_data}{A matrix of data to be heat mapped. Needs column and
rownames.}

\item{annotation_row}{A data frame of the annotation variable(s). Names must
match with input_data. If NULL returns normal pheatmap.}

\item{sort_col}{The name of the column to sort input_data and
annotation_row by when heatmapping. If pheatmap is instructed to sort_rows
this has no impact.}

\item{main}{The default title for the heatmap}

\item{...}{The usual inputs for pheatmap.}
}
\value{
An annotated pheatmap from the pheatmap package
}
\description{
Returns and prints an annotated pheatmap sorted by a given
column of the annotation data frame with clustering occuring within the
levels of said column
}
