% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{phi_prior}
\alias{phi_prior}
\title{Phi prior}
\usage{
phi_prior(matrix_data)
}
\arguments{
\item{matrix_data}{A matrix of data.}
}
\value{
A list of vectors of the proportion of each level across all of
matrix_data.
}
\description{
Generates a prior for the phi vector for each variable for the
Dirichlet distribution
}
