% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{prepare_cat_data}
\alias{prepare_cat_data}
\title{Prepare cat data}
\usage{
prepare_cat_data(data)
}
\arguments{
\item{data}{A data frame or matrix of categorical data}
}
\value{
A matrix of integers where the lowest class fval
}
\description{
Converts categorical data to numerical format appropriate for
analysis (i.e. integers with a null class of 0)
}
