% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SampleCategoryProbabilities}
\alias{SampleCategoryProbabilities}
\title{Sample the probabilities for each category across all clusters for each 
covariate}
\usage{
SampleCategoryProbabilities(data, class_probs, phi_prior, cluster_labels,
  cat_count, n_clust, n_col, class_start)
}
\arguments{
\item{data}{Matrix of categorical data}

\item{class_probs}{Field of matrices (one for each variable) of category
probability for each cluster.}

\item{phi_prior}{Prior on class_probs.}

\item{cluster_labels}{Current labelling within dataset.}

\item{cat_count}{Vector of the number of categories within each variable.}

\item{n_clust}{The number of clusters present.}

\item{n_col}{The number of variables present.}

\item{class_start}{0 or 1; allows labelling to begin at 0 or 1 (C++ 
convention versus that of R).}
}
\value{
Field of matrices of sampled class weights for each cluster within 
each variable.
}
\description{
Sample the probabilities for each category across all clusters for each 
covariate
}
