% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SampleDirichletPosterior}
\alias{SampleDirichletPosterior}
\title{Sample parameters for a dirichlet distribution (normally for the clusters).}
\usage{
SampleDirichletPosterior(concentration_0, cluster_labels, n_clusters,
  count_from)
}
\arguments{
\item{concentration_0}{Pior on the concentration parameter.}

\item{cluster_labels}{Current clustering.}

\item{n_clusters}{The number of clusters present.}

\item{count_from}{0 or 1; to allow for C++ or R counting.}
}
\value{
Vector sampled from Dirichlet posterior.
}
\description{
Sample parameters for a dirichlet distribution (normally for the clusters).
}
