\name{tagmmdi-package}
\alias{tagmmdi-package}
\alias{tagmmdi}
\docType{package}
\title{\packageTitle{tagmmdi}}
\description{\packageDescription{tagmmdi}}
\section{Package Content}{\packageIndices{tagmmdi}}
\author{\packageAuthor{tagmmdi}}
\section{Maintainer}{\packageMaintainer{tagmmdi}}
\keyword{package}
