% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{thinning_warning}
\alias{thinning_warning}
\title{Thinning warning}
\usage{
thinning_warning(thinning, num_iter, burn)
}
\value{
None.
}
\description{
Generates an error or warning if the relative values of
thinning, num_iter and burn (the arguments controlling the number of recorded
MCMC samples) are strange.
}
