% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{cluster_weight_prior}
\alias{cluster_weight_prior}
\title{Cluster weight prior}
\usage{
cluster_weight_prior(train_data, outlier = FALSE)
}
\description{
Produces a vector of prior weights for clusters
}
